requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto==2.40.0
celery==3.1.23
crc16==0.1.1
defusedxml==0.4.1
Django==1.9.7
django-compat==1.0.13
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-debug-toolbar==1.7
django-filter==0.13.0
django-formtools==1.0
django-ses-backend==0.1.1
django-storages==1.5.0
django-user-sessions==1.3.0
djangorestframework==3.4.6
ed25519==1.4
gunicorn==19.5.0
httplib2==0.9.2
kombu==3.0.35
Markdown==2.6.6
msgpack-python==0.4.7
oauthlib==1.1.2
phonenumbers==7.4.0
Pillow==3.2.0
psycopg2==2.6.2
PyJWT==1.4.0
pysha3==0.3
PySocks==1.5.6
pytz==2016.4
PyYAML==3.11
raven==6.0.0
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
sseclient==0.0.12
toml==0.9.2
twilio==5.4.0
factory_boy==2.8.1
django-oauth-toolkit==0.11.0
django-anymail[sendgrid]==0.8
django-celery-email==1.1.5
two1==3.10.4
